BytesLocation: make the backing ByteBuffer mutable
We will need this in the BytesHandle, because it might try to write too
much data to the buffer... and then what do you do? Well, you allocate a
new, larger buffer, and copy the contents from the old buffer into that.

Unfortunately, this has the confusing side effect of making the
BytesHandle's backing BytesLocation no longer wrap the originally given
ByteBuffer/byte[], but that's tough -- when you keep writing to a buffer
past its capacity, something has to give.