Do runtime type checking in Typed.supports method
There are situations where we cannot solely rely on the compiler.
For example, with DataHandle plugins, the generic parameter L is
heterogeneous, so we need to actually check the type of the data.

I tried to make this the default implementation of the Typed interface
itself, but ran into problems with calling super.supports in downstream
classes; apparently, you cannot directly reference default interface
methods by writing e.g. Typed.super.supports(...)? Strange.